Move controller to correct place
