+++ src/routes/api/2014/images.ts
@@ -1,6 +1,6 @@
 import * as express from 'express'
 
-import ImageController from '@/controllers/imageController'
+import ImageController from '@/controllers/api/imageController'
 
 const router = express.Router()
